refactor : reuse issue creator & reuse marker attribute key 'elementId'
